Add checks on number of allowed delegates
Add both server and client side checks on the number of
allowed delegates. Note that the algorithm for computing
the number of allowed delegates is not correct and is a
temporary placeholder.

+++ Controller/CouncilDelegatesController.php
@@ -27,34 +27,72 @@ class CouncilDelegatesController extends StandardController {
 
     $councilDelegates = $this->CouncilDelegate->find('all', $args);
 
+    // Query for active CO Person Roles for this COU and include
+    // PrimaryName for rendering the form.
+    $coPersonRoleModel = ClassRegistry::init('CoPersonRole');
+
+    $args = array();
+    $args['conditions']['CoPersonRole.cou_id'] = $couId;
+    $args['conditions']['CoPersonRole.status'] = StatusEnum::Active;
+    $args['contain']['CoPerson'] = 'PrimaryName';
+
+    $coPersonRoles = $coPersonRoleModel->find('all', $args);
+    usort($coPersonRoles, array($this, "coPersonPrimaryNameCmp"));
+
+    // Compute the number of allowed council delegates.
+    // TODO This is not the correct algorithm.
+    $groupWorkContribution = count($coPersonRoles);
+    $allowedDelegateNumber = intval(ceil($groupWorkContribution/5.0));
+
     // Process incoming POST CouncilDelegate data.
     if($this->request->is('post')) {
+      $selectedCount = 0;
       foreach($this->data['CouncilDelegate']['rows'] as $d) {
-        // Reset model state between save/delete calls.
-        $this->CouncilDelegate->clear();
-
-        // CO Person is current delegate and selected to remain a delegate.
-        if(!empty($d['id']) && $d['delegate'] == 1) {
-          continue;
+        if($d['delegate'] == 1) {
+          $selectedCount++;
         }
+      }
 
-        // CO Person is current delegate and selected to be removed so delete.
-        if(!empty($d['id']) && $d['delegate'] == 0) {
-          if(!$this->CouncilDelegate->delete($d['id'])) {
-            // TODO Flash error here.
+      // If the number of selected delegates is greater than the number of
+      // allowed delegates reject the POST and set the flash, else process
+      // the delegate changes.
+      if($selectedCount > $allowedDelegateNumber) {
+        $this->Flash->set(_txt('pl.ligo_council.error.delegate.count.high', array($allowedDelegateNumber)), array('key' => 'error'));
+      } else {
+        foreach($this->data['CouncilDelegate']['rows'] as $d) {
+          // Reset model state between save/delete calls.
+          $this->CouncilDelegate->clear();
+
+          // CO Person is current delegate and selected to remain a delegate.
+          if(!empty($d['id']) && $d['delegate'] == 1) {
+            continue;
           }
-          continue;
-        }
 
-        // CO Person is not current delegate and selected to become delegate.
-        if(empty($d['id']) && $d['delegate'] == 1) {
-          $newDelegate['CouncilDelegate'] = $d;
-          if(!$this->CouncilDelegate->save($newDelegate)) {
-            // TODO Flash error here.
+          // CO Person is current delegate and selected to be removed so delete.
+          if(!empty($d['id']) && $d['delegate'] == 0) {
+            if(!$this->CouncilDelegate->delete($d['id'])) {
+              // TODO Flash error here.
+            }
+            continue;
           }
+
+          // CO Person is not current delegate and selected to become delegate.
+          if(empty($d['id']) && $d['delegate'] == 1) {
+            $newDelegate['CouncilDelegate'] = $d;
+            if(!$this->CouncilDelegate->save($newDelegate)) {
+              // TODO Flash error here.
+            }
+          }
+
+          //TODO History Records
+        }
+        
+        if($selectedCount < $allowedDelegateNumber) {
+          $this->Flash->set(_txt('pl.ligo_council.info.delegate.count.low', array($allowedDelegateNumber)), array('key' => 'information'));
+        } else {
+          $this->Flash->set(_txt('pl.ligo_council.success.delegate.count.updated'), array('key' => 'success'));
         }
 
-        //TODO History Records
       }
 
       // Redirect back to index to render updated delegate information.
@@ -71,19 +109,9 @@ class CouncilDelegatesController extends StandardController {
 
     $this->set('council_delegates', $councilDelegates);
 
-    // Query for active CO Person Roles for this COU and include
-    // PrimaryName for rendering the form.
-    $coPersonRoleModel = ClassRegistry::init('CoPersonRole');
-
-    $args = array();
-    $args['conditions']['CoPersonRole.cou_id'] = $couId;
-    $args['conditions']['CoPersonRole.status'] = StatusEnum::Active;
-    $args['contain']['CoPerson'] = 'PrimaryName';
-
-    $coPersonRoles = $coPersonRoleModel->find('all', $args);
-    usort($coPersonRoles, array($this, "coPersonPrimaryNameCmp"));
-
     $this->set('co_person_roles', $coPersonRoles);
+
+    $this->set('allowed_delegate_number', $allowedDelegateNumber);
   }
 
   /*
